datapath/tables: Add fields to routes
The desired routes we will have for route reconciliation have the MTU
and Type fields which the routing table currently does not store.
In order to be able to compare observed and desired routes, add these
fields to the observed route objects.

Also add string methods for the scope and type field to allow for better
human readable table output.

+++ pkg/datapath/tables/node_address.go
@@ -350,7 +350,7 @@ func (n *nodeAddressController) reconcile() (<-chan struct{}, <-chan struct{}) {
 		// We are only interested in local routes, which are used for IPs that are assigned
 		// to the host (e.g. on GCE). These routes are in the local table, have host scope,
 		// and have no source address.
-		if route.Scope != uint8(RT_SCOPE_HOST) || route.Src.IsValid() {
+		if route.Scope != RT_SCOPE_HOST || route.Src.IsValid() {
 			continue
 		}
